Switch Keycloak Docker image from Docker Hub to ECR
We have been using ECR in the main TDR accounts for a few weeks, and it
has been working well, so switch the Keycloak Sandbox over to ECR too.

This requires you to push a Keycloak image to the Sandbox account, but
this is more flexible than when we were using Docker Hub.